Added: Seaweed Script Font Modified: Folder names
Added Seaweed Script font to the asset folder, which is 100% free. The Ocean one I'm currently using is not. The Seaweed Script font is for pop-ups like "+1 coin". I need to find other fonts for other text, then remove the Ocean one. Modified the folder names to start with "!" so that they are organised easier for me to find.